get possible fields while selecting a form + dynamic fields to TS
