+++ hat/assets/js/apps/Iaso/domains/instances/config.js
@@ -3,17 +3,18 @@ import orderBy from 'lodash/orderBy';
 
 import { INSTANCE_METAS_FIELDS } from './constants';
 import MESSAGES from './messages';
-import ActionTableColumnComponent from "./components/ActionTableColumnComponent";
+import ActionTableColumnComponent from './components/ActionTableColumnComponent';
 
-export const actionTableColumn = (formatMessage = () => ({}), user, dispatch) => {
-
+export const actionTableColumn = (formatMessage = () => ({}), user) => {
     return {
         Header: formatMessage(MESSAGES.actions),
         accessor: 'actions',
         resizable: false,
         sortable: false,
         width: 150,
-        Cell: settings => <ActionTableColumnComponent settings={settings} user={user}/>
+        Cell: settings => (
+            <ActionTableColumnComponent settings={settings} user={user} />
+        ),
     };
 };
