+++ hat/assets/js/apps/Iaso/domains/instances/types/visibleColumns.ts
@@ -1,0 +1,7 @@
+export type VisibleColumn = {
+    active: boolean;
+    disabled: boolean;
+    key: string;
+    label: string;
+    meta: boolean;
+};
